changed uint256 to uint128

+++ ethereum/contracts/RewardSource.sol
@@ -4,5 +4,5 @@ pragma experimental ABIEncoderV2;
 
 // Something that can reward a relayer
 interface RewardSource {
-    function reward(address payable feePayer, uint256 _amount) external;
+    function reward(address payable feePayer, uint128 _amount) external;
 }
